Move up .plugins.plugin.
